+++ src/entry/dependenciesNumber/dependenciesSimplifyConstant.generated.ts
@@ -13,6 +13,7 @@ import { SymbolNodeDependencies } from './dependenciesSymbolNode.generated.js'
 import { isBoundedDependencies } from './dependenciesIsBounded.generated.js'
 import { matrixDependencies } from './dependenciesMatrix.generated.js'
 import { typedDependencies } from './dependenciesTyped.generated.js'
+import { parseDependencies } from './dependenciesParse.generated.js'
 import { createSimplifyConstant } from '../../factoriesNumber.js'
 
 export const simplifyConstantDependencies: Record<string, unknown> = {
@@ -27,5 +28,6 @@ export const simplifyConstantDependencies: Record<string, unknown> = {
   isBoundedDependencies,
   matrixDependencies,
   typedDependencies,
+  parseDependencies,
   createSimplifyConstant
 }