+++ src/function/arithmetic/gcd.ts
@@ -157,7 +157,7 @@ export const createGcd = /* #__PURE__ */ factory(
             if (is1d(array)) {
               return self(...array)
             }
-            throw new Error('gcd() supports only 1d matrices!')
+            throw new ArgumentsError('gcd', 1, 1)
           }
         ),
         Matrix: typed.referToSelf(